Orbot integration
This patch went through several iterations, all with the goal of having
a liveliness to the toggle, so that when enabled, connections are
immediately proxied.

First, separate client factory classes were used, based on the pref.
However, activities seem to reuse clients for their lifetimes, so only
unexplored aspects on the app got proxied right away.

Since OkHttpClients respect a system proxy, a second approach was to
just dynamically update that.  Unfortunately, open connections in the
client's pool aren't evicted when the proxy settings change.

Thus, forcing a restart on pref change, though undesirable, seemed like
the most expedient method of achieving the goal.

Note that the preference to "useTor" is only exposed if Orbot is
installed.

Bug: T163747
Change-Id: Ida57827743b6472d12e9b2e5ed1cb507972c354e